test(libcnb): colocate owner tests
